- for the service layer create interfaces - create @Audit annotation for the services to leverage custom audit implementation
