+++ core/src/main/java/com/netflix/conductor/annotations/Audit.java
@@ -1,30 +1,34 @@
-/*
+/**
  * Copyright 2018 Netflix, Inc.
- * <p>
+ *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * <p>
- * http://www.apache.org/licenses/LICENSE-2.0
- * <p>
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.netflix.conductor.service;
+/**
+ * 
+ */
+package com.netflix.conductor.annotations;
 
-import java.util.List;
+import java.lang.annotation.Retention;
+import java.lang.annotation.Target;
 
-public interface WorkflowBulkService {
-   void pauseWorkflow(List<String> workflowIds);
+import static java.lang.annotation.ElementType.TYPE;
+import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
-   void resumeWorkflow(List<String> workflowIds);
+/**
+ * Mark service for custom audit implementation
+ */
+@Target({TYPE})
+@Retention(RUNTIME)
+public @interface Audit {
 
-   void restart(List<String> workflowIds, boolean useLatestDefinitions);
-
-   void retry(List<String> workflowIds);
-
-   void terminate(List<String> workflowIds, String reason);
 }
