+++ core/src/main/java/com/netflix/conductor/service/AdminService.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 Netflix, Inc.
+ * Copyright 2018 Netflix, Inc.
  * <p>
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,71 +15,26 @@
  */
 package com.netflix.conductor.service;
 
-import com.netflix.conductor.annotations.Trace;
 import com.netflix.conductor.common.metadata.tasks.Task;
-import com.netflix.conductor.core.config.Configuration;
-import com.netflix.conductor.core.execution.WorkflowExecutor;
-import com.netflix.conductor.dao.QueueDAO;
-import com.netflix.conductor.service.utils.ServiceUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-import javax.inject.Inject;
-import javax.inject.Singleton;
-import java.io.InputStream;
 import java.util.List;
 import java.util.Map;
-import java.util.Properties;
 
-/**
- * @author fjhaveri
- *
- */
-@Singleton
-@Trace
-public class AdminService {
+public interface AdminService {
 
-    private static Logger LOGGER = LoggerFactory.getLogger(AdminService.class);
-
-    private final Configuration config;
-
-    private final ExecutionService executionService;
-
-    private final QueueDAO queueDAO;
-
-    private String version;
-
-    private String buildDate;
-
-    @Inject
-    public AdminService(Configuration config, ExecutionService executionService, QueueDAO queueDAO) {
-        this.config = config;
-        this.executionService = executionService;
-        this.queueDAO = queueDAO;
-        this.version = "UNKNOWN";
-        this.buildDate = "UNKNOWN";
-
-        try {
-            InputStream propertiesIs = this.getClass().getClassLoader().getResourceAsStream("META-INF/conductor-core.properties");
-            Properties prop = new Properties();
-            prop.load(propertiesIs);
-            this.version = prop.getProperty("Implementation-Version");
-            this.buildDate = prop.getProperty("Build-Date");
-        } catch (Exception e) {
-            LOGGER.error(e.getMessage(), e);
-        }
-    }
+    /**
+     * Queue up all the running workflows for sweep.
+     *
+     * @param workflowId Id of the workflow
+     * @return the id of the workflow instance that can be use for tracking.
+     */
+    String requeueSweep(String workflowId);
 
     /**
      * Get all the configuration parameters.
      * @return all the configuration parameters.
      */
-    public Map<String, Object> getAllConfig() {
-        Map<String, Object> map = config.getAll();
-        map.put("version", version);
-        map.put("buildDate", buildDate);
-        return map;
-    }
+    Map<String, Object> getAllConfig();
 
     /**
      * Get the list of pending tasks for a given task type.
@@ -89,24 +44,5 @@ public class AdminService {
      * @param count Number of entries
      * @return list of pending {@link Task}
      */
-    public List<Task> getListOfPendingTask(String taskType, Integer start, Integer count) {
-        ServiceUtils.checkNotNullOrEmpty(taskType, "TaskType cannot be null or empty.");
-        List<Task> tasks = executionService.getPendingTasksForTaskType(taskType);
-        int total = start + count;
-        total = (tasks.size() > total) ? total : tasks.size();
-        if (start > tasks.size()) start = tasks.size();
-        return tasks.subList(start, total);
-    }
-
-    /**
-     * Queue up all the running workflows for sweep.
-     *
-     * @param workflowId Id of the workflow
-     * @return the id of the workflow instance that can be use for tracking.
-     */
-    public String requeueSweep(String workflowId) {
-        ServiceUtils.checkNotNullOrEmpty(workflowId, "WorkflowId cannot be null or empty.");
-        boolean pushed = queueDAO.pushIfNotExists(WorkflowExecutor.DECIDER_QUEUE, workflowId, config.getSweepFrequency());
-        return pushed + "." + workflowId;
-    }
+    List<Task> getListOfPendingTask(String taskType, Integer start, Integer count);
 }
